+++ App.js
@@ -1,70 +1,20 @@
-import React, { useEffect, useState } from 'react';
-import { StatusBar } from 'expo-status-bar';
-import { ActivityIndicator, StyleSheet, FlatList, Text, View, Image } from 'react-native';
-import { REACT_APP_BASE_URL } from '@env';
+import React from 'react';
+import { NavigationContainer } from '@react-navigation/native';
+import Recipes from './screens/recipes';
+import Recipe from './screens/recipe';
+import { createNativeStackNavigator } from '@react-navigation/native-stack';
+
+const Stack = createNativeStackNavigator();
 
 export default function App() {
-  const [isLoading, setLoading] = useState(true);
-  const [data, setData] = useState([]);
-
-  const getRecipes = async () => {
-    try {
-      const response = await fetch(`${REACT_APP_BASE_URL}`);
-      const json = await response.json();
-      setData(json);
-    } catch (error) {
-      console.error(error);
-    } finally {
-      setLoading(false);
-    }
-  }
-
-  useEffect(() => {
-    getRecipes();
-  }, []);
+  
 
   return (
-    <View style={styles.container}>
-      <Image
-        style={styles.logo}
-        source={require('./assets/LP_Retro_Logo.jpg')}
-      />
-      <StatusBar style="auto" />
-      {isLoading ? <ActivityIndicator size="large" color="#00ff00" /> : (
-        <>
-          <Text style={styles.header}>Recipes</Text>
-          <FlatList
-            data={data}
-            keyExtractor={({ id }, index) => id}
-            renderItem={({ item }) => (
-              <Text style={styles.name}>{item.name}</Text>
-            )}
-          />
-        </>
-      )}
-    </View>
+    <NavigationContainer>
+      <Stack.Navigator initialRouteName="Lolly's Recipes">
+        <Stack.Screen name="Lolly's Recipes" component={Recipes} />
+        <Stack.Screen name="Recipe Details" component={Recipe} />
+      </Stack.Navigator>
+    </NavigationContainer>
   );
 }
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    backgroundColor: '#fff',
-    alignItems: 'center',
-    justifyContent: 'center',
-  },
-  header: {
-    textAlign:'center',
-    color: '#ff1493',
-    fontWeight: 'bold',
-    fontSize: 30,
-    marginBottom:16,
-  },
-  logo: {
-    width: 200,
-    height: 200,
-  },
-  name: {
-    fontSize:16,
-  }
-});
